projects/5/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.5
django-cors-headers==3.13.0
django-grappelli==3.0.3
djangorestframework==3.13.1
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
Pillow==9.1.1
psycopg2==2.9.3
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PySide6==6.3.1
PySide6-Addons==6.3.1
PySide6-Essentials==6.3.1
pytz==2022.1
requests==2.28.1
shiboken6==6.3.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.10
yarl==1.7.2
